Move the plugin to the proper place
